+++ dflib/src/main/java/org/dflib/pivot/PivotBuilder.java
@@ -39,21 +39,41 @@ public class PivotBuilder {
     }
 
     /**
-     * Use values from "columnName" to create pivoted table columns. For meaningful results "columnName" should contain
+     * Use values from "columnName" to create pivoted table columns. For meaningful results, "columnName" should contain
      * "categorical" data.
+     *
+     * @since 1.0.0-M22
      */
-    public PivotBuilder columns(String columnName) {
+    public PivotBuilder cols(String columnName) {
         this.columnForColumns = validateColumn(columnName);
         return this;
     }
 
     /**
-     * Use values from "columnPos" to create pivoted table columns. For meaningful results "columnPos" should contain
-     * "categorical" data.
+     * @deprecated in favor of {@link #cols(String)}
      */
-    public PivotBuilder columns(int columnPos) {
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public PivotBuilder columns(String columnName) {
+        return cols(columnName);
+    }
+
+    /**
+     * Use values from "columnPos" to create pivoted table columns. For meaningful results, "columnPos" should contain
+     * "categorical" data.
+     *
+     * @since 1.0.0-M22
+     */
+    public PivotBuilder cols(int columnPos) {
         this.columnForColumns = validateColumn(columnPos);
         return this;
+    }
+
+    /**
+     * @deprecated in favor of {@link #cols(int)}
+     */
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public PivotBuilder columns(int columnPos) {
+        return cols(columnPos);
     }
 
     /**
@@ -75,22 +95,41 @@ public class PivotBuilder {
     /**
      * Executes pivot transform, using values from the provided column name to populate the resulting DataFame.
      * There must be no more than one value for each pivot row and column combination, or an exception will be thrown.
-     * For datasets where multiple values are present, use a flavor with value aggregator -
-     * {@link #values(int, Exp)}.
+     * For datasets where multiple values are present, use a flavor with value aggregator - {@link #vals(int, Exp)}.
+     *
+     * @since 1.0.0-M22
      */
-    public DataFrame values(String columnName) {
+    public DataFrame vals(String columnName) {
         int pos = validateColumn(columnName);
         return doPivot(pos, oneValueAgg);
+    }
+
+
+    /**
+     * @deprecated in favor of {@link #vals(String)}
+     */
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public DataFrame values(String columnName) {
+        return vals(columnName);
     }
 
     /**
      * Executes pivot transform, using values from the provided column name to populate the resulting DataFame.
      * There must be no more than one value for each pivot row and column combination, or an exception will be thrown.
-     * For datasets where multiple values are present, use a flavor with value aggregator -
-     * {@link #values(int, Exp)}.
+     * For datasets where multiple values are present, use a flavor with value aggregator - {@link #vals(int, Exp)}.
+     *
+     * @since 1.0.0-M22
      */
-    public DataFrame values(int columnPos) {
+    public DataFrame vals(int columnPos) {
         return doPivot(columnPos, oneValueAgg);
+    }
+
+    /**
+     * @deprecated in favor of {@link #vals(int)}
+     */
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public DataFrame values(int columnPos) {
+        return vals(columnPos);
     }
 
     /**
@@ -99,23 +138,39 @@ public class PivotBuilder {
      * <code>$decimal(0).sum()</code>. Notice that the column name or index can be anything, as the evaluation happens
      * against individual columns, not DataFrame.
      *
-     * @see Exp
+     * @since 1.0.0-M22
      */
-    public <T> DataFrame values(String columnName, Exp<T> valuesAggregator) {
+    public <T> DataFrame vals(String columnName, Exp<T> valuesAggregator) {
         int pos = validateColumn(columnName);
         return doPivot(pos, valuesAggregator);
     }
 
     /**
+     * @deprecated in favor of {@link #vals(String, Exp)}
+     */
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public <T> DataFrame values(String columnName, Exp<T> valuesAggregator) {
+        return vals(columnName, valuesAggregator);
+    }
+
+    /**
      * Executes pivot transform, using values from the provided column name to populate the resulting DataFame. Values
      * with matching pivot row and column are aggregated with the provided aggregator. Aggregator may look like this:
      * <code>$decimal(0).sum()</code>. Notice that the column name or index can be anything, as the evaluation happens
      * against individual columns, not DataFrame.
      *
-     * @see Exp
+     * @since 1.0.0-M22
      */
-    public <T> DataFrame values(int columnPos, Exp<T> valuesAggregator) {
+    public <T> DataFrame vals(int columnPos, Exp<T> valuesAggregator) {
         return doPivot(columnPos, valuesAggregator);
+    }
+
+    /**
+     * @deprecated in favor of {@link #vals(String, Exp)}
+     */
+    @Deprecated(since = "1.0.0-M22", forRemoval = true)
+    public <T> DataFrame values(int columnPos, Exp<T> valuesAggregator) {
+        return values(columnPos, valuesAggregator);
     }
 
     protected <T> DataFrame doPivot(int columnPos, Exp<T> valuesAggregator) {